Add hook ability animation for warrior
