+++ client/next-js/consts/spellCosts.json
@@ -27,7 +27,7 @@
   "shadow-leap": 25
   ,"warbringer": 30
   ,"savage-blow": 20
-  ,"hamstring": 40
+  ,"hook": 40
   ,"bladestorm": 30
   ,"berserk": 30
   ,"bloodthirst": 0
