+++ client/next-js/skills/warrior/index.js
@@ -1,6 +1,6 @@
 export { meta as warbringer } from './warbringer';
 export { meta as savageBlow } from './savageBlow';
-export { meta as hamstring } from './hamstring';
+export { meta as hook } from './hook';
 export { meta as bladestorm } from './bladestorm';
 export { meta as berserk } from './berserk';
 export { meta as bloodthirst } from './bloodthirst';
